tests/requirements.txt
selenium
python-owasp-zap-v2.4
urllib3
requests
